feat: Refactor event processing code
