+++ crates/walrus-service/src/event.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Walrus Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+//! Event related modules.
+
+pub mod event_blob;
+#[cfg(feature = "client")]
+pub mod event_blob_downloader;
+pub mod event_processor;
+pub mod events;
