+++ crates/walrus-service/src/event/event_processor.rs
@@ -1,0 +1,15 @@
+// Copyright (c) Walrus Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+//! Event processor for processing events from the Sui network.
+
+pub mod bootstrap;
+pub mod catchup;
+pub mod checkpoint;
+pub mod client;
+pub mod config;
+pub mod db;
+pub mod metrics;
+pub mod package_store;
+pub mod processor;
+pub mod runtime;
